[HOPSWORKS-1617] Add (#98)
* fixes for multi-worker clusters

+++ hopsworks-installer.sh
@@ -36,9 +36,9 @@ NON_INTERACT=0
 SCRIPTNAME=`basename $0`
 AVAILABLE_MEMORY=$(free -g | grep Mem | awk '{ print $2 }')
 AVAILABLE_DISK=$(df -h | grep '/$' | awk '{ print $4 }')
-AVAILABLE_MNT=$(df -h | grep '/mnt$' | awk '{ print $4 }')
+# Azure mounts /mnt/resource - AWS/GCP mount /mnt
+AVAILABLE_MNT=$(df -h | grep '/mnt' | awk '{ print $4 }')
 AVAILABLE_CPUS=$(cat /proc/cpuinfo | grep '^processor' | wc -l)
-AVAILABLE_GPUS=$(lspci | grep -i nvidia | wc -l)
 IP=$(hostname -I | awk '{ print $1 }')
 HOSTNAME=$(hostname -f)
 DISTRO=
@@ -59,6 +59,7 @@ REVERSE_DNS=1
 CLOUD=
 GCP_NVME=0
 RM_CLASS=
+RM_WORKER=
 ENTERPRISE=0
 KUBERNETES=0
 DOWNLOAD=
@@ -75,7 +76,6 @@ if [ $? -ne 0 ] ; then
    sudo yum install pciutils -y > /dev/null
 fi    
 AVAILABLE_GPUS=$(sudo lspci | grep -i nvidia | wc -l)
-
 unset_gpus()
 {
 RM_CLASS="hops:
@@ -156,7 +156,7 @@ splash_screen()
   echo "You appear to have following setup on this host:"
   echo "* available memory: $AVAILABLE_MEMORY"
   echo "* available disk space (on '/' root partition): $AVAILABLE_DISK"
-  echo "* available disk space (on '/mnt' partition): $AVAILABLE_MNT"  
+  echo "* available disk space (under '/mnt' partition): $AVAILABLE_MNT"  
   echo "* available CPUs: $AVAILABLE_CPUS"
   echo "* available GPUS: $AVAILABLE_GPUS"  
   echo "* your ip is: $IP"
@@ -495,7 +495,7 @@ add_worker()
       echo "Cannot add worker node, as you need to be able to ssh into it using your public key"
       echo ""
       echo ""
-      echo -n "You can setup passwordless SSH to setup to ${USER}@${WORKER_IP} by entering the password."
+      echo "You can setup passwordless SSH to setup to ${USER}@${WORKER_IP} by entering the password."
       echo "Running ssh-copy-id.... "
       ssh-copy-id -i ${HOME}/.ssh/id_rsa.pub ${USER}@${WORKER_IP}
       if [ $? -ne 0 ] ; then
@@ -507,13 +507,12 @@ add_worker()
    WORKER_DISK=$(ssh -t -o StrictHostKeyChecking=no $WORKER_IP "df -h | grep '/\$' | awk '{ print \$4 }'") 
    WORKER_CPUS=$(ssh -t -o StrictHostKeyChecking=no $WORKER_IP "cat /proc/cpuinfo | grep '^processor' | wc -l")
 
-   NUM_GBS=$(expr $AVAILABLE_MEMORY - 2)
-   NUM_CPUS=$(expr $AVAILABLE_CPUS - 1)
+   NUM_GBS=$(expr $WORKER_MEM - 2)
+   NUM_CPUS=$(expr $WORKER_CPUS - 1)
    
    echo "Amount of disk space available on root partition ('/'): $WORKER_DISK"
-   echo "Amount of memory available on worker: $WORKER_MEM"
-   printf 'Please enter the amout of memory in this worker to be used (GBs)'
-   echo -n ". Default is $NUM_GBS: "
+   echo "Amount of memory available on this worker: $WORKER_MEM"
+   printf "Please enter the amout of memory in this worker to be used (GBs): "
    read GBS
    if [ "$GBS" == "" ] ; then
        GBS=$NUM_GBS
@@ -521,8 +520,7 @@ add_worker()
    
    MBS=$(expr $GBS \* 1024)
    echo "Amount of CPUs available on worker: $WORKER_CPUS"
-   printf 'Please enter the number of CPUs in this worker to be used'
-   echo -n ". Default is $NUM_CPUS: "   
+   printf "Please enter the number of CPUs in this worker to be used: "   
    read CPUS
    if [ "$CPUS" == "" ] ; then
        CPUS=$NUM_CPUS
@@ -532,28 +530,43 @@ add_worker()
        ssh -t -o StrictHostKeyChecking=no $WORKER_IP "sudo yum install pciutils -y"
    fi
 
+   WORKER_MNT=$(ssh -t -o StrictHostKeyChecking=no $WORKER_IP "sudo df -h | grep '/mnt' | awk '{ print $4 }'")
+   re='^[0-9]+$'
+
+   
    if [ "$CLOUD" == "azure" ] ; then
-       NSLOOKUP=$(ssh -t -o StrictHostKeyChecking=no $WORKER_IP "nslookup $WORKER_IP")
-       SUSPECTED_HOSTNAME=$(echo $NSLOOKUP | grep name | awk {' print $4 '} | grep -v 'internal.cloudapp.net')
+       NSLOOKUP=$(ssh -t -o StrictHostKeyChecking=no $WORKER_IP "nslookup $WORKER_IP | grep name | grep -v 'internal.cloudapp.net'")
+       SUSPECTED_HOSTNAME=$(echo $NSLOOKUP | awk {' print $4 '})
        SUSPECTED_HOSTNAME=${SUSPECTED_HOSTNAME::-1}
        echo ""
        echo "On Azure, you need to add every worker to the same Private DNS Zone, and note the hostname you set in Azure."
        echo "We suspect the private DNS hostname is:"
        echo "    $SUSPECTED_HOSTNAME"
        echo ""
-       printf 'Please enter the private DNS hostname for this worker (default:'
-       echo -n " $SUSPECTED_HOSTNAME):"
+       printf  "Please enter the private DNS hostname for this worker (default: $SUSPECTED_HOSTNAME): "
        read PRIVATE_HOSTNAME
        if [ "$PRIVATE_HOSTNAME" == "" ] ; then
 	   PRIVATE_HOSTNAME=$SUSPECTED_HOSTNAME
        fi
        ssh -t -o StrictHostKeyChecking=no $WORKER_IP "sudo hostname $PRIVATE_HOSTNAME"
+
+
+       if [[ $WORKER_MNT =~ $re ]] ; then
+          ssh -t -o StrictHostKeyChecking=no $WORKER_IP "sudo rm -rf /srv/hops; sudo mkdir -p /mnt/resource/hops; sudo ln -s /mnt/resource/hops /srv/hops"
+       fi
+   else
+       if [[ $WORKER_MNT =~ $re ]] ; then
+          ssh -t -o StrictHostKeyChecking=no $WORKER_IP "sudo rm -rf /srv/hops; sudo mkdir -p /mnt/hops; sudo ln -s /mnt/hops /srv/hops"
+       fi
    fi       
    
    WORKER_GPUS=$(ssh -t -o StrictHostKeyChecking=no $WORKER_IP "sudo lspci | grep -i nvidia | wc -l")
-   if [ "$WORKER_GPUS" == "" ] ; then
-     WORKER_GPUS=0
-   fi
+   echo "GPUs found on worker: $WORKER_GPUS"
+
+   # re='^[0-9]+$'
+   # if ! [[ $WORKER_GPUS =~ $re ]] ; then
+   #   WORKER_GPUS="0"
+   # fi
    echo ""
    echo "Number of GPUs found on worker: $WORKER_GPUS"
    echo ""
@@ -575,9 +588,16 @@ add_worker()
    fi
 
    if [[ "$WORKER_GPUS" > "0" ]] ; then
-       set_gpus
+      RM_WORKER="cuda:
+        accept_nvidia_download_terms: true
+      hops:
+        capacity: 
+          resource_calculator_class: org.apache.hadoop.yarn.util.resource.DominantResourceCalculatorGPU
+        yarn:
+          gpus: '*'"
    else
-       unset_gpus
+      RM_WORKER="hops:
+        yarn:"
    fi
 echo "  
   datanode${WORKER_ID}:
@@ -585,7 +605,7 @@ echo "
     baremetal:
       ip: ${WORKER_IP}
     attrs:
-      $RM_CLASS
+      $RM_WORKER
           vcores: $CPUS
           memory_mbs: $MBS
     recipes:
@@ -643,9 +663,17 @@ install_dir()
    mnt="${AVAILABLE_MNT//G}"
    
    if [ "$mnt" != "" ] && [ $mnt -gt $root ] ; then
-       sudo mkdir -p /mnt/hops
-       sudo rm -rf /srv/hops
-       sudo ln -s /mnt/hops /srv/hops
+       # Azure mounts disks here: /mnt/reosurce
+       if [ "$CLOUD" == "azure" ] ; then
+	   sudo mkdir -p /mnt/resource/hops
+	   sudo rm -rf /srv/hops
+	   sudo ln -s /mnt/resource/hops /srv/hops
+       else
+       # AWS/GCP mount disks here: /mnt
+	   sudo mkdir -p /mnt/hops
+	   sudo rm -rf /srv/hops
+	   sudo ln -s /mnt/hops /srv/hops
+       fi
    fi
 }
 
@@ -701,7 +729,7 @@ purge_local()
    sudo rm -rf ~/.karamel   
    sudo rm -rf /tmp/chef-solo/cookbooks
    echo "Purging old installation..."      
-   sudo rm -rf /srv/hops/*   
+   sudo rm -rf /srv/hops
 }
 
 ###################################################################################################
@@ -1008,7 +1036,7 @@ else
     if [ $? -ne 0 ] ; then
 	echo ""
 	echo "It appears you need a sudo password for this account."
-        echo -n "Enter the sudo password for $USER: "
+        echo "Enter the sudo password for $USER: "
 	read -s passwd
         SUDO_PWD="-passwd $passwd"
 	echo ""
@@ -1045,7 +1073,7 @@ else
     perl -pi -e "s/__TLS__/$TLS/" $YML_FILE
     if [ $ENTERPRISE -eq 1 ] ; then
 	echo ""
-        echo -n "Enter the URL to download the Enterprise Binaries from: "
+        printf "Enter the URL to download the Enterprise Binaries from: "
 	read DOWNLOAD_URL
 	DOWNLOAD_URL=${DOWNLOAD_URL//\./\\\.}
 	DOWNLOAD_URL=${DOWNLOAD_URL//\//\\\/}
